refactor: create Header component
