+++ .babelrc
@@ -3,7 +3,7 @@
     [
       "babel-plugin-styled-components",
       {
-        "displayName": false
+        "displayName": true
       }
     ]
   ]
